// apps/frontend/src/api/openapi.d.ts
export interface components {
    schemas: {
// ...
        JobCopyRequest: {
            /** Name */
            name?: string | null;
        };
        /** JobCreateRequest */
        JobCreateRequest: {
            /** Name */
            name: string;
            /** Description */
            description?: string | null;
            /** Config */
            config: {
                [key: string]: unknown;
            };
        };
        /** JobResponse */
        JobResponse: {
            /** Id */
            id: number;
            /** Name */
            name: string;
            /** Description */
            description: string | null;
            status: components["schemas"]["JobStatus"];
            /** Latest Config Id */
            latest_config_id: number | null;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /**
             * Updated At
             * Format: date-time
             */
            updated_at: string;
        };
        /**
         * JobStatus
         * @enum {string}
         */
        JobStatus: "draft" | "ready" | "archived";
        /** JobUpdateRequest */
        JobUpdateRequest: {
            /** Name */
            name?: string | null;
            /** Description */
            description?: string | null;
            status?: components["schemas"]["JobStatus"] | null;
        };
// ...
        RunResponse: {
            /** Id */
            id: string;
            /** Job Id */
            job_id: number;
            /** Config Id */
            config_id: number;
            status: components["schemas"]["RunStatus"];
            /** Process Id */
            process_id: number | null;
            /** Command */
            command: string;
            /** Started At */
            started_at: string | null;
            /** Ended At */
            ended_at: string | null;
            /** Exit Code */
            exit_code: number | null;
            /** Error Message */
            error_message: string | null;
            /** Job Name */
            job_name: string | null;
            /** Job Description */
            job_description: string | null;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /**
             * Updated At
             * Format: date-time
             */
            updated_at: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}